Убрала лишнее, добавила комментарии
